+++ benchmarks/runner/CHANGELOG.md
@@ -1,5 +1,11 @@
 # claudexor-benchmark-runner
 
+## 3.9.3
+
+### Patch Changes
+
+- @claudexor/util@3.9.3
+
 ## 3.9.2
 
 ### Patch Changes