+++ packages/acp-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @claudexor/acp-server
 
+## 3.9.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @claudexor/schema@3.9.3
+  - @claudexor/util@3.9.3
+
 ## 3.9.2
 
 ### Patch Changes