+++ packages/claudexor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # claudexor
 
+## 3.9.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @claudexor/cli@3.9.3
+
 ## 3.9.2
 
 ### Patch Changes