+++ packages/journal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @claudexor/journal
 
+## 3.9.3
+
+### Patch Changes
+
+- @claudexor/util@3.9.3
+
 ## 3.9.2
 
 ### Patch Changes